Refactor with package directory
