+++ wikipron/_version.py
@@ -1,0 +1,2 @@
+# This module is executed by setup.py.
+__version__ = "0.1.1"
