+++ wikipron/languagecodes.py
@@ -1,0 +1,16 @@
+# Map from a ISO 639 code to its non-ISO 639 Wiktionary language name.
+# ISO 639-3: https://iso639-3.sil.org/sites/iso639-3/files/downloads/iso-639-3.tab  # noqa: E501
+# Wiktionary: https://en.wiktionary.org/wiki/Category:Terms_with_IPA_pronunciation_by_language  # noqa: E501
+# TODO: Expand this as needed to cover additional languages.
+LANGUAGE_CODES = {
+    # Greek. Would have been "Modern Greek (1453-)" in ISO 639.
+    "el": "Greek",
+    "ell": "Greek",
+    "gre": "Greek",
+    "greek": "Greek",
+    # Slovene. Would have been "Slovenian" in ISO 639.
+    "sl": "Slovene",
+    "slv": "Slovene",
+    "slovene": "Slovene",
+    "slovenian": "Slovene",
+}
